+++ src/odbc/unittests/utf8.c
@@ -50,10 +50,13 @@ TestBinding(int minimun)
 		SQLLEN s1_len, s2_len;
 		unsigned int len;
 
-		len = minimun ? (strlen(strings_hex[p-strings]) - 2) /4 : 40;
+		len = minimun ? ((int) strlen(strings_hex[p-strings]) - 2) / 4
+		    : 40;
 		CHKBindParameter(2, SQL_PARAM_INPUT, SQL_C_CHAR,
 			SQL_WCHAR, len, 0, (void *) p[0], 0, &s1_len, "S");
-		len = minimun ? (strlen(strings_hex[p+1-strings]) - 2) /4 : 40;
+		len = minimun
+		    ? ((int) strlen(strings_hex[p+1-strings]) - 2) / 4
+		    : 40;
 		/* FIXME this with SQL_VARCHAR produce wrong protocol data */
 		CHKBindParameter(3, SQL_PARAM_INPUT, SQL_C_CHAR,
 			SQL_WVARCHAR, len, 0, (void *) p[1], 0, &s2_len, "S");
